debug: auth api build err debugging
