+++ app/api/login-kakao/check-nickname/route.ts
@@ -1,4 +1,5 @@
-import { NextRequest, NextResponse } from 'next/server'
+import type { NextRequest } from 'next/server'
+import { NextResponse } from 'next/server'
 import { isNicknameTaken } from '@/lib/repositories/user'
 
 export async function GET(req: NextRequest) {
