+++ app/api/login-kakao/start/route.ts
@@ -1,4 +1,5 @@
-import { NextRequest, NextResponse } from 'next/server'
+import type { NextRequest } from 'next/server'
+import { NextResponse } from 'next/server'
 
 // Delegate to NextAuth sign-in flow for Kakao
 export async function GET(req: NextRequest) {
