+++ app/api/profile/delete/route.ts
@@ -1,10 +1,10 @@
 import { NextResponse } from 'next/server'
 import { getServerSession } from 'next-auth'
-import { authOptions } from '@/app/api/auth/[...nextauth]/route'
+import { authOptions } from '@/app/api/auth/[...nextauth]/auth-options'
 import prisma from '@/lib/prisma'
 
 export async function POST() {
-  const session = await getServerSession(authOptions as any)
+  const session = await getServerSession(authOptions)
   const userId = (session as any)?.userId as string | undefined
   if (!userId) {
     return NextResponse.json({ ok: false, error: 'unauthorized' }, { status: 401 })
